+++ src/main/java/ru/practicum/shareit/user/service/UserService.java
@@ -1,49 +1,16 @@
 package ru.practicum.shareit.user.service;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import ru.practicum.shareit.exception.DublicateException;
 import ru.practicum.shareit.user.model.User;
-import ru.practicum.shareit.user.storage.UserStorage;
 
 import java.util.List;
 
-@Service
-@RequiredArgsConstructor
-public class UserService {
-    private final UserStorage userStorage;
+public interface UserService {
 
-    public User createUser(User user) {
-        checkEmail(user);
-        return userStorage.createUser(user);
-    }
+     User createUser(User user);
 
-    public User getUserById(long id) {
-        return userStorage.getUserById(id);
-    }
+     User getUserById(long id);
+     List<User> getAllUsers();
+     User updateUser(User user);
 
-    public List<User> getAllUsers() {
-        return userStorage.getAllUsers();
-    }
-
-    public User updateUser(User user) {
-        checkEmail(user);
-        return userStorage.updateUser(user);
-    }
-
-    public void deleteUser(long id) {
-        userStorage.deleteUser(id);
-    }
-
-    private void checkEmail(User user) {
-        if (userStorage.getAllUsers().stream()
-                .anyMatch(
-                        stored -> stored.getEmail().equalsIgnoreCase(user.getEmail())
-                                && stored.getId() != user.getId()
-                )
-        ) {
-            throw new DublicateException("User with this email address mail " +
-                    user.getEmail() + " already exists");
-        }
-    }
+     void deleteUser(long id);
 }
