+++ src/main/java/ru/practicum/shareit/user/service/UserServiceImpl.java
@@ -1,0 +1,50 @@
+package ru.practicum.shareit.user.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import ru.practicum.shareit.exception.DublicateException;
+import ru.practicum.shareit.user.model.User;
+import ru.practicum.shareit.user.storage.UserStorage;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class UserServiceImpl implements UserService {
+
+    private final UserStorage userStorage ;
+
+    public User createUser(User user) {
+        checkEmail(user);
+        return userStorage.createUser(user);
+    }
+
+    public User getUserById(long id) {
+        return userStorage.getUserById(id);
+    }
+
+    public List<User> getAllUsers() {
+        return userStorage.getAllUsers();
+    }
+
+    public User updateUser(User user) {
+        checkEmail(user);
+        return userStorage.updateUser(user);
+    }
+
+    public void deleteUser(long id) {
+        userStorage.deleteUser(id);
+    }
+
+    private void checkEmail(User user) {
+        if (userStorage.getAllUsers().stream()
+                .anyMatch(
+                        stored -> stored.getEmail().equalsIgnoreCase(user.getEmail())
+                                && stored.getId() != user.getId()
+                )
+        ) {
+            throw new DublicateException("Пользователь с таким адресом Эл. почты " +
+                    user.getEmail() + " уже существует!");
+        }
+    }
+}
